# docs/src/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Message Vault HTTP API",
    "description": "HTTP API for a local Message Vault. Bearer session tokens come from login. API tokens come from Settings → Account.",
    "license": {
      "name": "Fair Core License 1.0 (ALv2 future)",
      "url": "https://github.com/bitrealm-io/message-vault/blob/main/LICENSE.md"
    },
    "version": "0.8.3"
  },
  "paths": {
...
    "/v1/conversations": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "summary": "Page through conversations with participants, message counts, and tags.\nOrdered by most recent activity unless `sort` and `order` say otherwise.",
        "operationId": "conversations_list_handler",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Conversation search; empty lists all non-trashed",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, default 40, max 500",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset, max 50000",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Order by `date` (last message, default) or `messages` (message count)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "`asc` or `desc` (default)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ConversationSummary"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/saved-searches": {
      "get": {
        "tags": [
          "Saved searches"
        ],
        "summary": "List the account's saved searches, A–Z.",
        "operationId": "saved_searches_list_handler",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Saved searches"
        ],
        "summary": "Create a saved search and return it with the updated list.",
        "operationId": "saved_searches_create_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/saved-searches/{id}": {
      "delete": {
        "tags": [
          "Saved searches"
        ],
        "summary": "Delete a saved search and return the updated list.",
        "description": "Deleting an import-created saved search removes the shortcut only. The\n`vault_imports` row it pointed at is the account's permanent record of that\nrun and is never touched here.",
        "operationId": "saved_searches_delete_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Saved search id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchDeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Saved searches"
        ],
        "summary": "Replace a saved search's name and query, and return the updated list.",
        "operationId": "saved_searches_update_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Saved search id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationParticipant": {
        "type": "object",
        "description": "One participant with display name and handle.",
        "required": [
          "handle",
          "service"
        ],
        "properties": {
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Linked vault contact id, when the handle is linked. Matches the `id`\nevery other contact shape uses, so a caller can compare the two without\nconverting either."
          },
          "handle": {
            "type": "string",
            "description": "Raw handle value (phone, email, or username)."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name from the import or the vault contact."
          },
          "name_alias": {
            "type": [
              "string",
              "null"
            ],
            "description": "Per service+identity alias from `contact_handles` when linked."
          },
          "service": {
            "type": "string",
            "description": "Platform service, e.g. `imessage`."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every failure: one sentence, with the HTTP status carrying the meaning.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable description of the failure."
          }
        }
      },
...
      "Page_ConversationSummary": {
        "type": "object",
        "description": "One page of a list.",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Conversation row for the list: participants, counts, tags.",
              "required": [
                "id",
                "participants",
                "message_count",
                "last_message_at",
                "service",
                "is_group",
                "tags"
              ],
              "properties": {
                "date_range_end": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Timestamp of the conversation's last message."
                },
                "date_range_start": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Timestamp of the conversation's first message."
                },
                "id": {
                  "type": "integer",
                  "format": "int64",
                  "description": "The conversation's id; search for it as `in:#<id>`."
                },
                "is_group": {
                  "type": "boolean",
                  "description": "True for group conversations."
                },
                "label": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Group label from the export, when present."
                },
                "last_message_at": {
                  "type": "string",
                  "description": "Timestamp of the last message."
                },
                "message_count": {
                  "type": "integer",
                  "format": "int64",
                  "description": "Messages in the conversation (excluding hidden duplicates).",
                  "minimum": 0
                },
                "participants": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConversationParticipant"
                  },
                  "description": "Participants with names and handles."
                },
                "service": {
                  "type": "string",
                  "description": "Platform service of the conversation, e.g. `imessage`."
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Message tags on this conversation."
                }
              }
            },
            "description": "The rows on this page."
          },
          "limit": {
            "type": "integer",
            "description": "Page size used.",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "description": "Page offset used.",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Rows matching the query across every page.",
            "minimum": 0
          }
        }
      },
...
      "SavedSearch": {
        "type": "object",
        "description": "One row of `saved_searches`.",
        "required": [
          "id",
          "name",
          "query",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Saved search id, unique across the vault."
          },
          "kind": {
            "type": "string",
            "description": "`manual` or `import`."
          },
          "name": {
            "type": "string",
            "description": "Display name, unique per account."
          },
          "query": {
            "type": "string",
            "description": "Query string, run against the conversation list."
          }
        }
      },
      "SavedSearchBody": {
        "type": "object",
        "description": "A saved search's name and query.",
        "required": [
          "name",
          "query"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "query": {
            "type": "string"
          }
        }
      },
      "SavedSearchDeleteResponse": {
        "type": "object",
        "description": "The updated list after deletion.",
        "required": [
          "ok",
          "savedSearches"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "savedSearches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SavedSearch"
            }
          }
        }
      },
      "SavedSearchResponse": {
        "type": "object",
        "description": "The affected saved search plus the updated list.",
        "required": [
          "savedSearch",
          "savedSearches"
        ],
        "properties": {
          "savedSearch": {
            "$ref": "#/components/schemas/SavedSearch"
          },
          "savedSearches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SavedSearch"
            }
          }
        }
      },
      "SavedSearchesListResponse": {
        "type": "object",
        "description": "The account's saved searches, A–Z.",
        "required": [
          "savedSearches"
        ],
        "properties": {
          "savedSearches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SavedSearch"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}